requirements.txt
requests
python-dotenv
airtable-python-wrapper
instagrapi
Pillow